refactor(backend): extract payload analysis helpers (#430)
